test: resolve QA rejections for ontology validators coverage
- Removed unauthorized `tests/test_ontology_validators.py` file sprawl.
- Deleted redundant hardcoded tests duplicating existing property-based validations.
- Consolidated deterministic array sorting tests within the `tests/contracts/test_ontology_validators.py` file.
- Upgraded synthetic edge case tests for `ActiveInferenceContract` and `BrowserDOMState` to use parameterized Hypothesis fuzzing validators.
